requirements.txt
requests==2.32.5
smolagents==1.24.0
langfuse==4.0.1
python-dotenv==1.2.2
openinference-instrumentation-smolagents==0.1.24
e2b-code-interpreter>=1.0.3
ddgs>=9.0.0

# document_RAG (LlamaIndex + Chroma). If pip upgrades huggingface-hub to >=1.0, reinstall smolagents deps:
#   pip install "huggingface-hub>=0.31.2,<1.0.0"
llama-index>=0.10.0
llama-index-vector-stores-chroma
llama-index-embeddings-huggingface
llama-index-llms-huggingface-api
chromadb
sentence-transformers